Broke out manifest to it's own file
